# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Task Management backend server",
        "title": "heisenflow-System",
        "contact": {
            "name": "HeisenGo",
            "url": "https://github.com/HeisenGo"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
        "/boards": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new board for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "Create user board",
                "parameters": [
                    {
                        "description": "Board details",
                        "name": "board",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.CreateBoardReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "board: the created board details",
                        "schema": {
                            "$ref": "#/definitions/presenter.CreateBoardResponse"
                        }
                    },
                    "400": {
                        "description": "error: bad request, invalid board details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/boards/invite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invite a user to a board with a specified role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "Invite user to board",
                "parameters": [
                    {
                        "description": "Invitation details",
                        "name": "invite",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.InviteUserToBoard"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "invite: the details of the invitation",
                        "schema": {
                            "$ref": "#/definitions/presenter.InviteMemberResponse"
                        }
                    },
                    "400": {
                        "description": "error: bad request, invalid invitation details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "error: forbidden, permission denied to invite",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/boards/my-boards/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the boards associated with the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "Get user's boards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uesrID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "boards: paginated list of user's boards",
                        "schema": {
                            "$ref": "#/definitions/presenter.BoardUserResp"
                        }
                    },
                    "400": {
                        "description": "error: bad request, wrong claim type",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/boards/publics/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the public boards.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "Get public boards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "boards: paginated list of public boards",
                        "schema": {
                            "$ref": "#/definitions/presenter.BoardUserResp"
                        }
                    },
                    "400": {
                        "description": "error: bad request, wrong claim type",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/boards/{boardID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a full board by its ID for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "Get full board by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Board ID",
                        "name": "boardID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "board: the full board details",
                        "schema": {
                            "$ref": "#/definitions/presenter.FullBoardResp"
                        }
                    },
                    "400": {
                        "description": "error: bad request, wrong claim type or invalid board ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a specific board by its ID for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "Delete board",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Board ID",
                        "name": "boardID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request, invalid user claims, or board ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden, user does not have permission to delete the board",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "presenter.BoardColumnResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.BoardTaskResp"
                    }
                }
            }
        },
        "presenter.BoardTaskResp": {
            "type": "object",
            "properties": {
                "end_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "start_at": {
                    "type": "string"
                },
                "story_at": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "presenter.BoardUserResp": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "presenter.ColumnResponseItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                }
            }
        },
...
        "presenter.CreateBoardReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "myboard123"
                },
                "type": {
                    "type": "string",
                    "example": "private(public)"
                }
            }
        },
        "presenter.CreateBoardResponse": {
            "type": "object",
            "properties": {
                "board_id": {
                    "type": "string"
                },
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ColumnResponseItem"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "presenter.FullBoardResp": {
            "type": "object",
            "properties": {
                "board_id": {
                    "type": "string"
                },
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.BoardColumnResp"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.BoardUserResp"
                    }
                }
            }
        },
...
        "presenter.InviteMemberResponse": {
            "type": "object",
            "properties": {
                "board_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "user_board_role_id": {
                    "type": "string"
                }
            }
        },
        "presenter.InviteUserToBoard": {
            "type": "object",
            "properties": {
                "board_id": {
                    "type": "string",
                    "example": "aeec51f9-dde3-409d-9415-df771f5b8a62"
                },
                "email": {
                    "type": "string",
                    "example": "inviatee_email.com"
                },
                "role": {
                    "type": "string",
                    "example": "editor"
                },
                "user_board_role_id": {
                    "type": "string",
                    "example": "31e8d41b-a84e-41c6-9564-4e932fccf213"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Type 'Bearer' followed by a space and your JWT token.\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}